AG-9272 Move datum types to seriesTypes.ts
This avoids too many cyclical imports in tooltipManager.ts (which uses
the SeriesNodeDatum and ErrorBoardSeriesNodeDatum in the makeTooltipMeta
method).
