+++ packages/ag-charts-community/src/chart/series/hierarchy/hierarchySeries.ts
@@ -1,7 +1,7 @@
 import type { ModuleContext } from '../../../module/moduleContext';
 import type { PointLabelDatum } from '../../../util/labelPlacement';
-import type { SeriesNodeDatum } from '../series';
 import { Series, SeriesNodePickMode } from '../series';
+import type { SeriesNodeDatum } from '../seriesTypes';
 
 export abstract class HierarchySeries<S extends SeriesNodeDatum> extends Series<S> {
     constructor(moduleCtx: ModuleContext) {
