+++ packages/ag-charts-community/src/module-support.ts
@@ -35,6 +35,7 @@ export * from './chart/series/seriesEvents';
 export * from './chart/series/seriesLabelUtil';
 export * from './chart/series/seriesMarker';
 export * from './chart/series/seriesTooltip';
+export * from './chart/series/seriesTypes';
 export * from './chart/series/cartesian/cartesianSeries';
 export * from './chart/series/cartesian/barUtil';
 export * from './chart/series/cartesian/areaUtil';
